Handle stale fork default branches when straight pushing
My play-scala-macwire-di-example fork still has 2.6.x as the default
branch (the official template repo has 2.7.x as the default).  So in
order to checkout the target branch, while having both an origin remote
(mine) and an upstream remote (playframework), we have to first,
conditionally, create the branch, then check it out.

For some reason in git (both in JGit and in console git) there is no one
command for this, thus the use of "if".